Move Backend class to separate backend.py
This avoids the need for some modules to depend on telescope_state.py,
and keeps the sizes of modules more manageable.
