+++ katsdptelstate/memory.py
@@ -19,7 +19,7 @@ import re
 import logging
 
 from . import utils
-from .telescope_state import Backend
+from .backend import Backend
 from .errors import ImmutableKeyError
 from .rdb_utility import dump_string, dump_zset
 try:
